% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_theme.R
\name{st_theme}
\alias{st_theme}
\title{A ggplot theme for datain}
\usage{
st_theme()
}
\description{
This function creates a ggplot theme for datain report (inspired by the Agricultural Census 2023 theme)
}
\examples{
\dontrun{
   p + st_theme()
}

}
